Fix SignalR error that only occurs on Azure, Microsoft.Azure.SignalR.Common.AzureSignalRUnauthorizedException
https://www.pivotaltracker.com/story/show/177805073
