+++ azure-functions/modules/encodeMsClientPrincipalName.ts
