+++ svelte-app/src/modules/signalR.ts
@@ -3,6 +3,7 @@ import type OpenCrypto from 'opencrypto';
 import type { IGlobalFeedback } from '../stores/globalFeedback';
 import { unconditionalMessage, showUnconditionalMessage, globalFeedback } from '../stores/globalFeedback';
 import { api } from './api';
+import { encodeMsClientPrincipalName } from './encodeMsClientPrincipalName';
 import getDefaultFunctionsUrl from './getFunctionsUrl';
 import type { NegotiateRequest, NegotiateResponse } from './serverInterfaces';
 import { NegotiateResponseType } from './serverInterfaces';
@@ -11,8 +12,6 @@ import { sign } from './sign';
 const GET = 'GET';
 const POST = 'POST';
 const NOT_FOUND = -1;
-const LENGTH_OF_ALPHANUMERIC = 36;
-const STRING_START = 0;
 
 export async function connectSignalR(
     options: {
@@ -76,6 +75,8 @@ async function getSignalRConnectionInfo(
     const url = `${getDefaultFunctionsUrl()}api/negotiate`;
     const method = signingPrivateKey ? POST : GET;
     let body: NegotiateRequest;
+    const encodedMsClientPrincipalName =
+        encodeMsClientPrincipalName(xMsClientPrincipalName);
     const response = await api<NegotiateResponse>(
         signingPrivateKey
             ? {
@@ -91,12 +92,12 @@ async function getSignalRConnectionInfo(
                     signingKey: signingPrivateKey,
                     omitBody: true
                 }),
-                xMsClientPrincipalName
+                xMsClientPrincipalName: encodedMsClientPrincipalName
             }
             : {
                 method,
                 url,
-                xMsClientPrincipalName
+                xMsClientPrincipalName: encodedMsClientPrincipalName
             });
     switch (response.type) {
         case NegotiateResponseType.Success:
@@ -143,32 +144,4 @@ async function getSignalRConnectionInfo(
         }
     }
     return response;
-}
-
-const safeCharCodes = new Set<number>(
-    [ ...Array(LENGTH_OF_ALPHANUMERIC).keys() ] // 0-9 and a-z
-        .map(idx => idx.toString(LENGTH_OF_ALPHANUMERIC).charCodeAt(STRING_START))
-        .concat(
-            [ '-', '.', '_' ].map(
-                safeChar => safeChar.charCodeAt(STRING_START))
-        ));
-
-// Need to encode SignalR user names to prevent errors which only occur on Azure.
-// Oddly, the emulator had no such issues. Example problematic email: dbruck1+Admin@fau.edu
-// Error from Azure:
-// Microsoft.Azure.SignalR.Common.AzureSignalRUnauthorizedException : Authorization failed.
-// Make sure you provide the correct connection string and have access to the resource. Request Uri:
-export function encodeMsClientPrincipalName(xMsClientPrincipalName: string) : string {
-    return [ ...Array(xMsClientPrincipalName.length).keys() ].reduce(
-        (previousValue, currentValue) => {
-            const nextCharCode = xMsClientPrincipalName.charCodeAt(currentValue);
-            return safeCharCodes.has(nextCharCode)
-                ? previousValue + xMsClientPrincipalName[currentValue]
-                : previousValue + escapeCharCode(nextCharCode);
-        },
-        '');
-}
-
-function escapeCharCode(charCode: number) : string {
-    return `~${charCode}~`;
 }
